% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plaus_weight.R
\name{add_plaus_weight}
\alias{add_plaus_weight}
\title{add plausible weight (by scenario and week)}
\usage{
add_plaus_weight(proj, variant_takeover, keep_flags = FALSE)
}
\arguments{
\item{proj}{data.table containing projections to determine plausibility}

\item{variant_takeover}{vector with dates of each variant takeover (names of
vector are names of variant)}

\item{keep_flags}{logical TRUE to keep the "plaus_scenario" and \code{plaus_week}
flags generated by helper functions}
}
\value{
proj data.table containing projections with column for scenario plausibility
}
\description{
add plausible weight (by scenario and week)
}
