% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_trend.R
\name{calc_thresh}
\alias{calc_thresh}
\title{calculate threshold for "increasing" "flat" or "decreasing" classification
based on \% change values}
\usage{
calc_thresh(dat_change, pct_flat)
}
\arguments{
\item{dat_change}{data.table of data including \% change column}

\item{pct_flat}{double specifying the \% of data that should be classified as "flat"}
}
\value{
data.table with thresholds for increasing/decreasing by target
}
\description{
calculate threshold for "increasing" "flat" or "decreasing" classification
based on \% change values
}
