% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_trend.R
\name{calculate_pct_change}
\alias{calculate_pct_change}
\title{calculate the \% change in series of data points by some lag}
\usage{
calculate_pct_change(dat, lag, byvars = NULL)
}
\value{
data.table with data and \% change values
}
\description{
calculate the \% change in series of data points by some lag
}
