% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_trend.R
\name{classify_incdec}
\alias{classify_incdec}
\title{classify data as increasing or decreasing based on threshold}
\usage{
classify_incdec(dat_change, thresh)
}
\arguments{
\item{dat_change}{data.table of data including \% change column}

\item{thresh}{data.table with thresholds for increasing/decreasing by target}
}
\value{
data.table with data classified as increasing, flat or decreasing
}
\description{
classify data as increasing or decreasing based on threshold
}
