% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.R
\name{implement_all_exclusions}
\alias{implement_all_exclusions}
\title{implement exclusions to SMH projections}
\usage{
implement_all_exclusions(
  proj,
  NA_vals = TRUE,
  quantile01 = TRUE,
  non_public = TRUE,
  territories = TRUE,
  model_names = TRUE,
  non_monotonic = TRUE,
  outside_projperiod = TRUE,
  proj_period,
  wrong_round = TRUE,
  scenario_round,
  summarize_exclusions = FALSE,
  summarize_exclusions_path = "code/evaluation/data/exclusions.csv"
)
}
\arguments{
\item{proj}{data.table containing projections}

\item{NA_vals}{logical TRUE to exclude all NA values in \code{value} or \code{quantile} columns}

\item{quantile01}{logical TRUE to exclude quantiles 0 and 1}

\item{non_public}{logical TRUE to exclude non-public rounds (8 and 10)}

\item{territories}{logical TRUE to exclude territories (locations 60, 66, 69, 72, 74, 78)}

\item{model_names}{vector of model names to exclude from projections, otherwise NA to skip this exclusion}

\item{outside_projperiod}{logical TRUE to exclude projections outside of the official SMH projection period}

\item{proj_period}{data.frame containing all accepted target_end_dates (by round)}

\item{wrong_round}{logical TRUE to exclude projections that were submitted for scenarios from a different round}

\item{scenario_round}{data.frame containing all accepted scenarios (by round)}
}
\value{
data.table containing projections after exclusion
}
\description{
implement exclusions to SMH projections
}
