% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_null_data.R
\name{load_all_null_projs}
\alias{load_all_null_projs}
\title{load projections for all 3 null models: GAM (gam), COVID-19 Forecast Hub
4-week ahead ensemble (fh), COVID-19 Forecast Hub 4-week ahead baseline (naive)}
\usage{
load_all_null_projs(only_win_proj_period = TRUE)
}
\arguments{
\item{only_win_proj_period}{logical TRUE if FH projections will be loaded for
only dates within projection period (others are included in the file
for increasing/decreasing calculations)}
}
\description{
load projections for all 3 null models: GAM (gam), COVID-19 Forecast Hub
4-week ahead ensemble (fh), COVID-19 Forecast Hub 4-week ahead baseline (naive)
}
