% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_null_data.R
\name{load_null_fh}
\alias{load_null_fh}
\title{load projections for COVID-19 Forecast Hub 4-week ahead ensemble null model}
\usage{
load_null_fh(pp_flag = TRUE, p = "code/evaluation/data/FH_proj.csv")
}
\arguments{
\item{p}{string path to file containing null projections}

\item{only_win_proj_period}{logical TRUE if FH projections will be loaded for
only dates within projection period (others are included in the file
for increasing/decreasing calculations)}
}
\description{
load projections for COVID-19 Forecast Hub 4-week ahead ensemble null model
}
