% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_null_data.R
\name{load_null_gam}
\alias{load_null_gam}
\title{load projections for GAM null model}
\usage{
load_null_gam(p = "code/evaluation/data/GAM_proj.csv")
}
\arguments{
\item{p}{string path to file containing null projections}
}
\description{
load projections for GAM null model
}
