% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_null_data.R
\name{load_null_naive}
\alias{load_null_naive}
\title{load projections for COVID-19 Forecast Hub 4-week ahead baseline model}
\usage{
load_null_naive(p = "code/evaluation/data/naive_proj.csv")
}
\arguments{
\item{p}{string path to file containing null projections}
}
\description{
load projections for COVID-19 Forecast Hub 4-week ahead baseline model
}
