% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_scores.R
\name{load_scores}
\alias{load_scores}
\title{function to load all scores from separate \code{.csv} files into one object in R}
\usage{
load_scores(base_dir)
}
\arguments{
\item{base_dir}{string location of directory where separate \code{.csv} files are
stored}
}
\value{
data.table containing all scores
}
\description{
function to load all scores from separate \code{.csv} files into one object in R
}
