% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plaus_weight.R
\name{plaus_weight_by_scenario}
\alias{plaus_weight_by_scenario}
\title{add plausible weight by scenario}
\usage{
plaus_weight_by_scenario(
  p = "code/evaluation/data/MostPlausibleScenarios.csv",
  proj
)
}
\arguments{
\item{p}{string location of file that specifies plausible weight of each scenario}

\item{proj}{data.table containing projections to determine plausibility}
}
\value{
proj data.table containing projections with column for scenario plausibility
}
\description{
add plausible weight by scenario
}
