% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plaus_weight.R
\name{plaus_weight_by_week}
\alias{plaus_weight_by_week}
\title{add plausible weight by week}
\usage{
plaus_weight_by_week(variant_takeover, proj)
}
\arguments{
\item{variant_takeover}{vector with dates of each variant takeover (names of vector are names of variant)}

\item{proj}{data.table containing projections to determine plausibility}
}
\value{
proj data.table containing projections with column for scenario plausibility
}
\description{
add plausible weight by week
}
